\name{TESS-package}
\alias{TESS-package}
\alias{TESS}
\docType{package}
\title{Diversification rate estimation and fast simulation of reconstructed phylogenetic trees under tree-wide time-heterogeneous birth-death processes including mass-extinction events}
\description{Simulation of reconstructed phylogenetic trees under tree-wide time-heterogeneous birth-death processes and estimation of parameters under the same model. Speciation and extinction rates can be any function of time and mass-extinction events at specific times can be provided. Trees can be simulated either conditioned on the number of species, the time of the process, or both. Additionally, the likelihood equations are implemented for convenience and can be used for Maximum Likelihood (ML) estimation and Bayesian inference.}
\details{
\tabular{ll}{
Package: \tab TESS\cr
Type: \tab Package\cr
Version: \tab 2.0.0\cr
Date: \tab 2014-09-24\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{Sebastian Hoehna and Michael R. May

Maintainer: Sebastian Hoehna <Sebastian.Hoehna@gmail.com>
}
\references{S. Hoehna: Fast simulation of reconstructed phylogenies under global, time-dependent birth-death processes. 2013, Bioinformatics, 29:1367-1374 }
%\keyword{tree, phylogeny, simulation}
\seealso{
\code{ape}
%\code{\link[ape]{ape}}
}
