% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess.analysis.efbd.R
\name{tess.analysis.efbd}
\alias{tess.analysis.efbd}
\title{tess.analysis.efbd}
\usage{
tess.analysis.efbd(
  tree,
  initialSpeciationRate,
  initialExtinctionRate,
  initialFossilizationRate,
  empiricalHyperPriors = TRUE,
  empiricalHyperPriorInflation = 10,
  empiricalHyperPriorForm = c("lognormal", "normal", "gamma"),
  speciationRatePriorMean = 0,
  speciationRatePriorStDev = 1,
  extinctionRatePriorMean = 0,
  extinctionRatePriorStDev = 1,
  fossilizationRatePriorMean = 0,
  fossilizationRatePriorStDev = 1,
  initialSpeciationRateChangeTime = c(),
  initialExtinctionRateChangeTime = c(),
  initialFossilizationRateChangeTime = c(),
  estimateNumberSpeciationRateChanges = TRUE,
  estimateNumberExtinctionRateChanges = TRUE,
  estimateNumberFossilizationRateChanges = TRUE,
  numExpectedRateChanges = 2,
  samplingProbability = 1,
  samplingStrategy = "uniform",
  missingSpecies = c(),
  timesMissingSpecies = c(),
  tInitialMassExtinction = c(),
  pInitialMassExtinction = c(),
  pMassExtinctionPriorShape1 = 5,
  pMassExtinctionPriorShape2 = 95,
  estimateMassExtinctionTimes = TRUE,
  numExpectedMassExtinctions = 2,
  estimateNumberMassExtinctions = TRUE,
  MRCA = TRUE,
  CONDITION = "survival",
  BURNIN = 10000,
  MAX_ITERATIONS = 2e+05,
  THINNING = 100,
  OPTIMIZATION_FREQUENCY = 500,
  CONVERGENCE_FREQUENCY = 1000,
  MAX_TIME = Inf,
  MIN_ESS = 500,
  ADAPTIVE = TRUE,
  dir = "",
  priorOnly = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{tree}{the phylogeny}

\item{initialSpeciationRate}{The initial value of the speciation rate when the MCMC is started.}

\item{initialExtinctionRate}{The initial value of the extinction rate when the MCMC is started.}

\item{initialFossilizationRate}{The initial value of the fossilization rate when the MCMC is started.}

\item{empiricalHyperPriors}{.}

\item{empiricalHyperPriorInflation}{.}

\item{empiricalHyperPriorForm}{.}

\item{speciationRatePriorMean}{mean parameter for the lognormal prior on lambda}

\item{speciationRatePriorStDev}{standard deviation for the lognormal prior on lambda}

\item{extinctionRatePriorMean}{mean parameter for the lognormal prior on mu}

\item{extinctionRatePriorStDev}{standard deviation for the lognormal prior on mu}

\item{fossilizationRatePriorMean}{mean parameter for the lognormal prior on phi}

\item{fossilizationRatePriorStDev}{standard deviation for the lognormal prior on phi}

\item{initialSpeciationRateChangeTime}{The initial value of the time points when speciation rate-shifts occur.}

\item{initialExtinctionRateChangeTime}{The initial value of the time points when extinction rate-shifts occur.}

\item{initialFossilizationRateChangeTime}{The initial value of the time points when fossilization rate-shifts occur.}

\item{estimateNumberSpeciationRateChanges}{Do we estimate the number of rate shifts?}

\item{estimateNumberExtinctionRateChanges}{Do we estimate the number of rate shifts?}

\item{estimateNumberFossilizationRateChanges}{Do we estimate the number of rate shifts?}

\item{numExpectedRateChanges}{Expected number of rate changes which follow a Poisson process.}

\item{samplingProbability}{probability of uniform sampling at present}

\item{samplingStrategy}{Which strategy was used to obtain the samples (taxa). Options are: uniform|diversified|age}

\item{missingSpecies}{The number of species missed which originated in a given time interval (empirical taxon sampling)}

\item{timesMissingSpecies}{The times intervals of the missing species (empirical taxon sampling)}

\item{tInitialMassExtinction}{The initial value of the vector of times of the mass-extinction events.}

\item{pInitialMassExtinction}{The initial value of the vector of survival probabilities of the mass-extinction events.}

\item{pMassExtinctionPriorShape1}{The alpha (first shape) parameter of the Beta prior distribution for the survival probability of a mass-extinction event.}

\item{pMassExtinctionPriorShape2}{The beta (second shape) parameter of the Beta prior distribution for the survival probability of a mass-extinction event.}

\item{estimateMassExtinctionTimes}{Do we estimate the times of mass-extinction events?}

\item{numExpectedMassExtinctions}{Expected number of mass-extinction events which follow a Poisson process.}

\item{estimateNumberMassExtinctions}{Do we estimate the number of mass-extinction events?}

\item{MRCA}{does the tree start at the mrca?}

\item{CONDITION}{do we condition the process on nothing|survival|taxa?}

\item{BURNIN}{number of starting iterations to be dropped}

\item{MAX_ITERATIONS}{The maximum number of iteration of the MCMC.}

\item{THINNING}{The frequency how often samples are recorded during the MCMC.}

\item{OPTIMIZATION_FREQUENCY}{The frequency how often the MCMC moves are optimized}

\item{CONVERGENCE_FREQUENCY}{The frequency how often we check for convergence?}

\item{MAX_TIME}{The maximum time the MCMC is allowed to run in seconds.}

\item{MIN_ESS}{The minimum number of effective samples (ESS) to assume convergence.}

\item{ADAPTIVE}{Do we use auto-tuning of the MCMC moves?}

\item{dir}{The subdirectory in which the output will be stored.}

\item{priorOnly}{Do we sample from the prior only?}

\item{verbose}{Do we want to print the progress of the MCMC?}
}
\value{
nothing. outputs are written to the file system
}
\description{
Running an analysis on a given tree and estimating the diversification rates including rate-shifts and mass-extinction events.
}
\examples{
# we load the conifers as the test data set
data(conifers)

# for the conifers we know what the total number of species is
total <- 630
# thus, we can compute what the sampling fraction is
rho <- (conifers$Nnode+1)/total


# next, we specify the prior mean and standard deviation
# for the speciation and extinction rate
mu_lambda = 0.15
std_lambda = 0.02
mu_mu = 0.09
std_mu = 0.02
mu_phi = 0.10
std_phi = 0.02



# now we can run the entire analysis.
# note that a full analyses should be run much longer
tess.analysis.efbd( tree=conifers,
                    initialSpeciationRate = exp(mu_lambda),
                    initialExtinctionRate = exp(mu_mu),
                    initialFossilizationRate = exp(mu_phi),
                    empiricalHyperPriors = FALSE,
                    speciationRatePriorMean = mu_lambda,
                    speciationRatePriorStDev = std_lambda,
                    extinctionRatePriorMean = mu_mu,
                    extinctionRatePriorStDev = std_mu,
                    fossilizationRatePriorMean = mu_phi,
                    fossilizationRatePriorStDev = std_phi,
                    numExpectedRateChanges = 2,
                    samplingProbability = rho,
                    numExpectedMassExtinctions = 2,
                    BURNIN = 100,
                    MAX_ITERATIONS = 200,
                    THINNING = 10,
                    dir = "analysis_conifer")
}
