% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess.branching.times.R
\name{tess.branching.times}
\alias{tess.branching.times}
\title{Title}
\usage{
tess.branching.times(phy, tip.age.threshold = 1e-05)
}
\arguments{
\item{phy}{the phylogeny}

\item{tip.age.threshold}{threshold to be considered an extant tip (not fossil)}
}
\value{
a list with items c("sampled_ancestor", "fossil_tip", "age_parent", "age", "tip")
}
\description{
Title
}
\examples{
data(conifers)

nodes <- tess.branching.times(conifers)
}
