% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess.likelihood.bdstp.R
\name{tess.likelihood.bdstp}
\alias{tess.likelihood.bdstp}
\title{tess.likelihood.bdstp}
\usage{
tess.likelihood.bdstp(
  nodes,
  lambda,
  mu,
  phi,
  r,
  samplingProbability,
  samplingStrategy = "uniform",
  MRCA = TRUE,
  CONDITION = "survival",
  log = TRUE
)
}
\arguments{
\item{nodes}{node information from \code{tess.branching.times(phy)}}

\item{lambda}{speciation rate vector}

\item{mu}{extinction rate vector}

\item{phi}{serial sampling rate vector}

\item{r}{conditional probability that a lineage dies upon (serial) sampling}

\item{samplingProbability}{probability of uniform sampling at present}

\item{samplingStrategy}{Which strategy was used to obtain the samples (taxa). Options are: uniform|diversified|age}

\item{MRCA}{does the tree start at the mrca?}

\item{CONDITION}{do we condition the process on nothing|survival|sampleAtLeastOneLineage?}

\item{log}{likelhood in log-scale?}
}
\value{
log likelihood
}
\description{
Computation of the likelihood for a given tree under an episodic fossilized-birth-death model (i.e. piecewise constant rates).
}
\examples{
data(conifers)
nodes <- tess.branching.times(conifers)

logL <- tess.likelihood.bdstp(nodes,
                              lambda = 1.0, 
                              mu = 0.5,
                              phi = 0.1,
                              r = 0.0,
                              samplingProbability = 1.0)
}
