% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess.likelihood.ebdstp.R
\name{tess.likelihood.ebdstp}
\alias{tess.likelihood.ebdstp}
\title{tess.likelihood.ebdstp}
\usage{
tess.likelihood.ebdstp(
  nodes,
  lambda,
  mu,
  phi,
  r,
  samplingProbabilityAtPresent,
  rateChangeTimesLambda = c(),
  rateChangeTimesMu = c(),
  rateChangeTimesPhi = c(),
  rateChangeTimesR = c(),
  massDeathTimes = c(),
  massDeathProbabilities = c(),
  burstBirthTimes = c(),
  burstBirthProbabilities = c(),
  eventSamplingTimes = c(),
  eventSamplingProbabilities = c(),
  samplingStrategyAtPresent = "uniform",
  MRCA = TRUE,
  CONDITION = "survival",
  log = TRUE
)
}
\arguments{
\item{nodes}{node times from tess.branching.times}

\item{lambda}{birth (speciation or infection) rates}

\item{mu}{death (extinction or becoming non-infectious without treatment) rates}

\item{phi}{serial sampling (fossilization) rates}

\item{r}{treatment probability, Pr(death | sample) (does not apply to samples take at time 0)}

\item{samplingProbabilityAtPresent}{probability of uniform sampling at present}

\item{rateChangeTimesLambda}{times at which birth rates change}

\item{rateChangeTimesMu}{times at which death rates change}

\item{rateChangeTimesPhi}{times at which serial sampling rates change}

\item{rateChangeTimesR}{times at which treatment probabilities change}

\item{massDeathTimes}{time at which mass-deaths (mass-extinctions) happen}

\item{massDeathProbabilities}{probability of a lineage dying in a mass-death event}

\item{burstBirthTimes}{}

\item{burstBirthProbabilities}{}

\item{eventSamplingTimes}{time at which every lineage in the tree may be sampled}

\item{eventSamplingProbabilities}{probability of a lineage being sampled at an event sampling time}

\item{samplingStrategyAtPresent}{Which strategy was used to obtain the samples (taxa). Options are: uniform|diversified|age}

\item{MRCA}{does the tree start at the mrca?}

\item{CONDITION}{do we condition the process on nothing|survival|taxa?}

\item{log}{likelhood in log-scale?}
}
\value{
probability of the speciation times
}
\description{
Computation of the likelihood for a given tree under an episodic fossilized-birth-death model (i.e. piecewise constant rates).
}
