\name{TESS-package}
\alias{TESS-package}
\alias{TESS}
\docType{package}
\title{Fast simulation of reconstructed phylogenetic trees under time-dependent birth-death processes}
\description{TESS is a package for simulation of reconstructed phylogenetic trees under global, time-dependent birth-death processes. Speciation and extinction rates can be any function of time and mass-extinction events at specific times can be provided. Trees can be simulated either conditioned on the number of species, the time of the process, or both. Additionally, the likelihood equations are implemented for convenience and can be used for Maximum Likelihood (ML) estimation and Bayesian inference.}
\details{
\tabular{ll}{
Package: \tab TESS\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr
Date: \tab 2013-03-21\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{Sebastian Hoehna

Maintainer: Sebastian Hoehna <hoehna@math.su.se>
}
\references{S. Hoehna: Fast simulation of reconstructed phylogenies under global, time-dependent birth-death processes. Submitted. 2012.}
%\keyword{tree, phylogeny, simulation}
\seealso{
\code{ape}
%\code{\link[ape]{ape}}
}
